+++ CHANGELOG.md
@@ -5,9 +5,10 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 ### Changed
+- SHUTTER command renamed to SUSPEND
+- UNSHUTTER command renamed to UNSUSPEND
 - The database trigger for publishing on the event_log table is now added on application 
 startup and removed on application shut down 
-
 
 ## [2.3.1] - 2019-12-07
 ### Fixed